restructure: change frontend directory structure to more modular approach
