+++ frontend/src/modules/Categories/components/ConfirmDialog.tsx
